+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Mildly reformatted `README.md`
 - A development mode check for the `app.state.ip_pool`'s initialization (caused application failure in production only)
 - Applied `get_db` utility function in all applicable areas.
+- Unspecified IPv6 addresses are returned without hiding in `utilities.hide_ip`
 
 ### Fixed